+++ template/base/src/main/java/ru/surfstudio/standard/base/logger/RemoteLogger.kt
@@ -1,21 +1,5 @@
-/*
-  Copyright (c) 2018-present, SurfStudio LLC.
+package ru.surfstudio.standard.base.logger
 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
- */
-package ru.surfstudio.android.logger
-
-import ru.surfstudio.android.logger.remote_logging_strategies.RemoteLoggingStrategy
 import kotlin.reflect.KClass
 
 /**
@@ -26,11 +10,9 @@ object RemoteLogger {
     private val REMOTE_LOGGING_STRATEGIES = hashMapOf<KClass<*>, RemoteLoggingStrategy>()
 
     @JvmStatic
-
     fun getRemoteLoggingStrategies() = REMOTE_LOGGING_STRATEGIES
 
     @JvmStatic
-
     fun addRemoteLoggingStrategy(strategy: RemoteLoggingStrategy) =
             REMOTE_LOGGING_STRATEGIES.put(strategy::class, strategy)